changed the code to use the base XMLMetric class for xml-chrf
